Improve doc (#144)
* start splitting pyvinecopulib into submodules

* update examples and tests

* properly split the package and add basic unit tests

* Add plot infrastructure

* Update pyprohect.toml

* Switch back to having stats as part of main module

* Switch back to having stats as part of main module+1

* Add missing files from MANIFEST.in

* Simplify doc building/serving

* Ensure _pyvinecopulib appears as pyvinecopulib

* Simplify mkdocs

* add bivariate plots

* add vine plots

* py38-friendly type hints

* remove scipy dependency

* upgrade cibw

* try properly setting up Python with the right architecture

* make codacy happy

* determine python version

* remove python setup stuff

* drop windows 32 bits wheels

* drop all 32 bits wheels

* finalize docstring

* make codacy happy
