+++ MANIFEST.in
@@ -1,4 +1,4 @@
-include README.md LICENSE
+include README.md LICENSE pyproject.toml setup.py
 recursive-include src *.py *.hpp *.cpp
 recursive-include lib/wdm/include *.hpp
 recursive-include lib/vinecopulib/include *.hpp *.ipp
